+++ legacy_packages/chains/chains/7000.ts
@@ -5,11 +5,17 @@ export default {
   "explorers": [
     {
       "name": "ZetaChain Mainnet Explorer",
-      "url": "https://explorer.mainnet.zetachain.com",
+      "url": "https://explorer.zetachain.com",
       "standard": "none"
     }
   ],
   "faucets": [],
+  "icon": {
+    "url": "ipfs://QmP4Gnf4Lkp8q5LQVePNjAWxSqrw8vU2JAf7amcFz4vEUy",
+    "width": 712,
+    "height": 712,
+    "format": "png"
+  },
   "infoURL": "https://zetachain.com/docs/",
   "name": "ZetaChain Mainnet",
   "nativeCurrency": {