+++ packages/thirdweb/src/react/core/hooks/pay/useBuyHistory.ts
@@ -35,6 +35,7 @@ export type BuyHistoryQueryOptions = Omit<
  *  return <div> ... </div>
  * }
  * ```
+ * @buyCrypto
  */
 export function useBuyHistory(
   params?: BuyHistoryParams,