+++ packages/thirdweb/src/utils/signatures/signature-to-hex.ts
@@ -22,6 +22,8 @@ import { type Hex, hexToBigInt } from "../encoding/hex.js";
  *
  * console.log(hex);
  * // "0x6e100a352ec6ad1b70802290e18aeed190704973570f3b8ed42cb9808e2ea6bf4a90a229a244495b41890987806fcbd2d5d23fc0dbe5f5256c2613c039d76db81c"
+ * ```
+ * @utils
  */
 export function signatureToHex(signature: Signature): Hex {
   const { r, s, v } = signature;