+++ packages/thirdweb/src/wallets/eip5792/get-capabilities.ts
@@ -4,7 +4,7 @@ import { isInAppWallet } from "../in-app/core/wallet/index.js";
 import { getInjectedProvider } from "../injected/index.js";
 import type { Wallet } from "../interfaces/wallet.js";
 import { isSmartWallet } from "../smart/index.js";
-import { isWalletConnect } from "../wallet-connect/index.js";
+import { isWalletConnect } from "../wallet-connect/controller.js";
 import type { WalletId } from "../wallet-types.js";
 import type { WalletCapabilities, WalletCapabilitiesRecord } from "./types.js";
 
@@ -19,7 +19,7 @@ export type GetCapabilitiesResult = Prettify<
 /**
  * Get the capabilities of a wallet based on the [EIP-5792](https://eips.ethereum.org/EIPS/eip-5792) specification.
  *
- * Note: This function is dependent on the wallet's support for EIP-5792, but will not throw.
+ * @note This function is dependent on the wallet's support for EIP-5792, but will not throw.
  * **The returned object contains a `message` field detailing any issues with the wallet's support for EIP-5792.**
  *
  * @param {GetCapabilitiesOptions} options
@@ -34,7 +34,7 @@ export type GetCapabilitiesResult = Prettify<
  * const capabilities = await getCapabilities({ wallet });
  * ```
  *
- * @wallets
+ * @extension EIP5792
  */
 export async function getCapabilities<const ID extends WalletId = WalletId>({
   wallet,